Add reliable stop-and-wait single message send
In order to test the reliable delivery, this commit includes a few mock
functions that can be injected in place of the typical socket sendto(),
recvfrom(), and poll() functions.

This commit does not yet include support for reliable receipt, or
support for messages larger than can fit into a single packet.