Add kftp send file on top of reliable udp
kftp (Kirby's File Transfer Protocol) is a protocol on top of reliable
udp (rudp). This commit adds the send file functionality which will read
the contents of a file and send them out over rudp in a streaming
manner. This should allow for the transfer of large files without
excessive memory consumption.