(hack) Mock fwrite calls
The changes this commit introduces to mocks.c and mocks.h caused the
cmocka tests to return exit code 255. However, this error appeared to be
occuring after main exited, and did not appear when run under a
debugger. It is not clear to me why this is the case. To get around it
for the time being, this commit changes the makefile to run the tests
under lldb which has an exit code of 0