Modify rudp sender to ACK previous messages
This fixes an issue where a process would receive a message, ack it, and
then send a new message, but where the ack was lost. In this case, the
original sender would resend the message. However, the process had
already sent a new message and was therefore expecting only an ack to
its message. With this change, if the process receives messages while in
the rudp_send loop that have been recently acked, it will send another
ack.