Modify client to check for acks after a command
This addresses the case where the client completes a command, however
the server never received the final ack. This commit ensures that the
client will listen a little longer to send acks to any messages that the
server sends, until a message is received that does not need an ack.